Updating share sample, changed activation to no longer check hasStarted. This fixes Cortana.
