+++ Samples/ShareTarget/Views/MainPage.xaml
@@ -1,4 +1,4 @@
-<Page x:Class="ShareTarget.Views.MainPage"
+﻿<Page x:Class="ShareTarget.Views.MainPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:Core="using:Microsoft.Xaml.Interactions.Core"
@@ -13,82 +13,16 @@
         <vm:MainPageViewModel x:Name="ViewModel" />
     </Page.DataContext>
 
-    <Page.Resources>
-        <Style x:Name="Header" TargetType="TextBlock">
-            <Setter Property="FontWeight" Value="Bold" />
-        </Style>
-    </Page.Resources>
+    <RelativePanel Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <controls:PageHeader x:Name="pageHeader" Content="Share sample, Main page"
+                             RelativePanel.AlignLeftWithPanel="True" RelativePanel.AlignRightWithPanel="True"
+                             RelativePanel.AlignTopWithPanel="True" />
 
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition />
-        </Grid.RowDefinitions>
+        <TextBlock x:Name="mainTextBlock" Margin="16,12,0,0"
+                   RelativePanel.AlignLeftWithPanel="True" RelativePanel.Below="pageHeader"
+                   Text="Hello World" />
 
-        <controls:PageHeader Content="Share Target Page" />
+    </RelativePanel>
 
-        <Grid Grid.Row="1">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition />
-                <ColumnDefinition />
-                <ColumnDefinition />
-            </Grid.ColumnDefinitions>
-            <Image Grid.Column="0" Source="{Binding Logo}" />
-            <Image Grid.Column="1" Source="{Binding Thumbnail}" />
-            <Image Grid.Column="2" Source="{Binding Bitmap}" />
-        </Grid>
-
-        <StackPanel Grid.Row="2" Padding="12,4,0,0">
-
-            <TextBlock Style="{StaticResource Header}" Text="Content" />
-            <TextBlock Text="{Binding Content}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="QuickLink" />
-            <TextBlock Text="{Binding QuickLink}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Title" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.Title}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Description" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.Description}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ApplicationName" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.ApplicationName}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ApplicationListingUri" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.ApplicationListingUri}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.PackageFamilyName" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.PackageFamilyName}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.EnterpriseId" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.EnterpriseId}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ContentSourceWebLink" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.ContentSourceWebLink}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ContentSourceApplicationLink" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.ContentSourceApplicationLink}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.LogoBackgroundColor" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.LogoBackgroundColor}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Square30x30Logo" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.Square30x30Logo}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Thumbnail" />
-            <TextBlock Text="{Binding ShareOperation.Data.Properties.Thumbnail}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.RequestedOperation" />
-            <TextBlock Text="{Binding ShareOperation.Data.RequestedOperation}" />
-
-            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.AvailableFormats" />
-            <TextBlock Text="{Binding ShareOperation.Data.AvailableFormats}" />
-
-        </StackPanel>
-
-    </Grid>
 </Page>
-
