+++ Samples/ShareTarget/Views/SharePage.xaml
@@ -1,0 +1,96 @@
+<Page x:Class="ShareTarget.Views.SharePage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:Core="using:Microsoft.Xaml.Interactions.Core"
+      xmlns:Interactivity="using:Microsoft.Xaml.Interactivity"
+      xmlns:controls="using:Template10.Controls"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:local="using:ShareTarget.Views"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:vm="using:ShareTarget.ViewModels" mc:Ignorable="d">
+
+    <Page.DataContext>
+        <vm:SharePageViewModel x:Name="ViewModel" />
+    </Page.DataContext>
+
+    <Page.Resources>
+        <Style x:Name="Header" TargetType="TextBlock">
+            <Setter Property="FontWeight" Value="Bold" />
+        </Style>
+    </Page.Resources>
+
+    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition />
+        </Grid.RowDefinitions>
+
+        <controls:PageHeader Content="Share Target Page">
+            <AppBarButton Click="{x:Bind ViewModel.Complete}" Icon="Clear" Label="Complete" />
+        </controls:PageHeader>
+
+        <Grid Grid.Row="1">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition />
+                <ColumnDefinition />
+                <ColumnDefinition />
+            </Grid.ColumnDefinitions>
+            <Image Grid.Column="0" Source="{Binding Logo}" />
+            <Image Grid.Column="1" Source="{Binding Thumbnail}" />
+            <Image Grid.Column="2" Source="{Binding Bitmap}" />
+        </Grid>
+
+        <StackPanel Grid.Row="2" Padding="12,4,0,0">
+
+            <TextBlock Style="{StaticResource Header}" Text="Content" />
+            <TextBlock Text="{Binding Content}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="QuickLink" />
+            <TextBlock Text="{Binding QuickLink}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Title" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.Title}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Description" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.Description}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ApplicationName" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.ApplicationName}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ApplicationListingUri" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.ApplicationListingUri}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.PackageFamilyName" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.PackageFamilyName}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.EnterpriseId" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.EnterpriseId}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ContentSourceWebLink" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.ContentSourceWebLink}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.ContentSourceApplicationLink" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.ContentSourceApplicationLink}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.LogoBackgroundColor" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.LogoBackgroundColor}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Square30x30Logo" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.Square30x30Logo}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.Properties.Thumbnail" />
+            <TextBlock Text="{Binding ShareOperation.Data.Properties.Thumbnail}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.RequestedOperation" />
+            <TextBlock Text="{Binding ShareOperation.Data.RequestedOperation}" />
+
+            <TextBlock Style="{StaticResource Header}" Text="ShareOperation.Data.AvailableFormats" />
+            <TextBlock Text="{Binding ShareOperation.Data.AvailableFormats}" />
+
+        </StackPanel>
+
+    </Grid>
+</Page>
+
