+++ Template10.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.25029.0
+VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Templates (Project)", "Templates (Project)", "{E9D12E06-A958-4FD7-A2BE-CE590A1FDFDB}"
 EndProject
@@ -281,6 +281,8 @@ Global
 		{BE334D79-1D64-4638-B547-75400AA9A58B}.Debug|x64.Build.0 = Debug|x64
 		{BE334D79-1D64-4638-B547-75400AA9A58B}.Debug|x64.Deploy.0 = Debug|x64
 		{BE334D79-1D64-4638-B547-75400AA9A58B}.Debug|x86.ActiveCfg = Debug|x86
+		{BE334D79-1D64-4638-B547-75400AA9A58B}.Debug|x86.Build.0 = Debug|x86
+		{BE334D79-1D64-4638-B547-75400AA9A58B}.Debug|x86.Deploy.0 = Debug|x86
 		{BE334D79-1D64-4638-B547-75400AA9A58B}.Release|Any CPU.ActiveCfg = Release|x86
 		{BE334D79-1D64-4638-B547-75400AA9A58B}.Release|ARM.ActiveCfg = Release|ARM
 		{BE334D79-1D64-4638-B547-75400AA9A58B}.Release|ARM.Build.0 = Release|ARM
